requirements.txt
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
cssselect2==0.2.1
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.1
django-bot-crawler-blocker==1.0
django-import-export==1.2.0
djangorestframework==3.9.4
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
kiwisolver==1.1.0
lxml==4.4.1
matplotlib==3.1.1
mysqlclient==1.3.13
numpy==1.17.1
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.25.1
patsy==0.5.1
Pillow==6.1.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.2
reportlab==3.5.23
requests==2.22.0
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
statsmodels==0.10.1
tablib==0.13.0
tinycss2==1.0.2
urllib3==1.25.3
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
